Use a custom User model for Django authentication
Django uses a pre-defined User model for authentication. By using a
custom model we can add relevant and skip irrelevant fields while
retaining compatibility with Django auth. It is also recommended by
the Django docs [1] to define and register a custom User model, even
though it may be a simple placeholder using 'pass', as it is difficult
to migrate from the default should the need arise later.

We also need custom UserManager, UserAdmin, and admin forms that are
compatible with the custom model for Django authentication to work as
the built-in ones expect fields which are no longer present.

[1] https://bit.ly/2w71XVB