Get modifiable model fields based on user
The default behaviour in Django admin is to allow modification of all
model fields. This is not desired in our case as modifying e.g. when an
order was created or delivered, or removing all contents of an order,
will cause unexpected behaviour.

Fields non-superusers should not be allowed to modify are therefore set
as read-only. Any fields that it does not make sense for a non-superuser
to set, such as setting which organisation an event belongs to (as there
is only one option for non-superusers!), are hidden for the user.