Show a Snackbar after creating or undoing an order
If a user submits an order by mistake, they have to find the order in
the orders list, click it, wait for the modal to open, and then press
'delete'. By showing a Snackbar with an easily accessible undo action,
the user is notified that an order has been created and can quickly
undo the order.