Sync selected event with the input text
When a user selects an event from the events dropdown that event is
saved in state. If a user keeps typing text into the input field and
forces focus loss of the field without selecting an event, the text and
selected event will be out of sync. When a user clicks the button they
probably expect that an event will be created with the text in the
field. They will actually choose the event they previously clicked on
before typing additional text.

This commit fixes this undesired behaviour by ensuring the field text is
the source of truth.