Style the application's Snackbars
Set a global style for the Snackbars used in the application.

Snackbars have a default font of Roboto which is very different from the
font used in other places of the application. The element alignments are
also off, the text is too small, and the 'warning' variant has an
unreadable text colour.

Use the correct font, increase the font size, adjust the alignment, and
use readable colours.