Stage backend for real-time updates with WebSocket
Setup Django channels with Redis, add a WebSocket consumer and its
routing, and subscribe a callback to model save and delete signals.

There is one issue with using signals to push newly created orders to
clients. In order to populate an order with items, the order must first
be created and saved to the database. This will trigger the save signal,
and the order pushed to clients will not contain any items. The signal
is not re-sent when adding items (the order itself is not updated).

The order creation process therefore includes a workaround by creating
an order without sending a post_save signal and then manually sends it
once all order items are created.