Add a 'createorganisation' command to Django-admin (#65)
* Add a 'createorganisation' command to Django-admin

A superuser cannot be created unless there exists an organisation to
associate it with. However, an organisation cannot be created unless
there exists a superuser. This deadlock is properly fixed by adding
a command to create organisations without first creating a user.

This command is run using './manage.py createorganisation' or equivalent
syntax depending on your platform.

* Simplify organisation theme validation