Authenticate WebSockets using Django sessions
Replaces sending the authentication token in the URL query string with
Django sessions and thereby simplifying the authentication process.

The JavaScript WebSocket API does not allow developers to add headers to
the HTTP request. One of the common workarounds, which we used, is to
send the token in the URL query and use it to authenticate. As the
production servers logs HTTP requests, the tokens are exposed in them.